# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/games/wordle/start": {
      "post": {
        "description": "Starts or re-opens the team's Wordle session. The answer is never returned.",
        "operationId": "start",
        "parameters": [],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartGameBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WordleStartResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Round closed/paused, game disabled, or game already completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBodyDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "spm_access_token": []
          }
        ],
        "summary": "Start Wordle",
        "tags": [
          "wordle"
        ]
      }
    },
    "/games/wordle/guess": {
      "post": {
        "description": "Submits a 5-letter word guess and returns feedback only.",
        "operationId": "guess",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WordleGuessBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WordleGuessResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Not a valid 5-letter dictionary word.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBodyDto"
                }
              }
            }
          },
          "404": {
            "description": "No session for this team (start first).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBodyDto"
                }
              }
            }
          },
          "409": {
            "description": "Session busy or game already completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBodyDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "spm_access_token": []
          }
        ],
        "summary": "Submit a guess",
        "tags": [
          "wordle"
        ]
      }
    },
    "/games/wordle/finish": {
      "post": {
        "description": "Finalizes an expired session as TIMEOUT or returns the stored result for a terminal session.",
        "operationId": "finish",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinishBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameFinishResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Session still in progress and not expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBodyDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "spm_access_token": []
          }
        ],
        "summary": "Finish Wordle",
        "tags": [
          "wordle"
        ]
      }
    },
    "/games/shadow/start": {
      "post": {
        "description": "Starts or re-opens the team's Shadow session and returns the round's questions (answers hidden).",
        "operationId": "start",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShadowStartResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Round closed/paused, game disabled, already completed, or no questions configured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBodyDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "spm_access_token": []
          }
        ],
        "summary": "Start Shadow",
        "tags": [
          "shadow"
        ]
      }
    },
    "/games/shadow/answer": {
      "post": {
        "description": "Submits an answer for a question; the correct answer is revealed only after the question resolves.",
        "operationId": "answer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShadowAnswerBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShadowAnswerResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Unknown questionId or malformed body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBodyDto"
                }
              }
            }
          },
          "404": {
            "description": "No session for this team (start first).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBodyDto"
                }
              }
            }
          },
          "409": {
            "description": "Question already resolved, attempts exhausted, session busy, or game already completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBodyDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "spm_access_token": []
          }
        ],
        "summary": "Answer a question",
        "tags": [
          "shadow"
        ]
      }
    },
    "/games/shadow/finish": {
      "post": {
        "description": "Finalizes an expired session as TIMEOUT or returns the stored result for a terminal session.",
        "operationId": "finish",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinishBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameFinishResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Session still in progress and not expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBodyDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "spm_access_token": []
          }
        ],
        "summary": "Finish Shadow",
        "tags": [
          "shadow"
        ]
      }
    },
    "/games/cards/start": {
      "post": {
        "description": "Starts or re-opens the team's Cards session and returns the board positions (fronts hidden).",
        "operationId": "start",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardsStartResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Round closed/paused, game disabled, or game already completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBodyDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "spm_access_token": []
          }
        ],
        "summary": "Start Cards",
        "tags": [
          "cards"
        ]
      }
    },
    "/games/cards/move": {
      "post": {
        "description": "Flips one card; the front asset is revealed only on this flip. Two face-up cards resolve a match attempt.",
        "operationId": "move",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardsMoveBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardsMoveResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Unknown cardId or malformed body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBodyDto"
                }
              }
            }
          },
          "404": {
            "description": "No session for this team (start first).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBodyDto"
                }
              }
            }
          },
          "409": {
            "description": "Card already matched/revealed, session busy, or game already completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBodyDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "spm_access_token": []
          }
        ],
        "summary": "Flip a card",
        "tags": [
          "cards"
        ]
      }
    },
    "/games/cards/finish": {
      "post": {
        "description": "Finalizes an expired session as TIMEOUT or returns the stored result for a terminal session.",
        "operationId": "finish",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinishBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameFinishResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Session still in progress and not expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBodyDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "spm_access_token": []
          }
        ],
        "summary": "Finish Cards",
        "tags": [
          "cards"
        ]
      }
    },
...
  },
  "info": {
    "title": "Spider-Man IEEE Event Platform API",
    "description": "REST API for the Spider-Man themed IEEE event platform.\n\nAuth uses HTTP-only cookies: `spm_access_token` (player) and `spm_admin_token` (admin). Player and admin routes are marked with the matching cookie security scheme. See `docs/frontend.md` (integration guide) and `docs/api.md` (full reference).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api",
      "description": "The platform prefix — all routes live under /api."
    }
  ],
  "components": {
    "securitySchemes": {
      "spm_access_token": {
        "type": "apiKey",
        "in": "cookie",
        "name": "spm_access_token"
      },
      "spm_admin_token": {
        "type": "apiKey",
        "in": "cookie",
        "name": "spm_admin_token"
      }
    },
    "schemas": {
...
      "ApiErrorBodyDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 409
          },
          "error": {
            "type": "string",
            "example": "Conflict"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "description": "Human-readable error message(s)."
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      },
...
      "StartGameBodyDto": {
        "type": "object",
        "properties": {
          "clientActionId": {
            "type": "string",
            "description": "Idempotency key (8-64 chars, alphanumeric + `-`).",
            "example": "start-uuid"
          }
        }
      },
      "WordleStartResponseDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "Game session id."
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "Server-side expiry — the countdown is visual only."
          },
          "attemptsAllowed": {
            "type": "number",
            "example": 6
          },
          "wordLength": {
            "type": "number",
            "example": 5
          }
        },
        "required": [
          "sessionId",
          "expiresAt",
          "attemptsAllowed",
          "wordLength"
        ]
      },
      "WordleGuessBodyDto": {
        "type": "object",
        "properties": {
          "guess": {
            "type": "string",
            "description": "Exactly 5 letters, a real dictionary word.",
            "example": "crane"
          },
          "clientActionId": {
            "type": "string",
            "description": "Idempotency key (8-64 chars, alphanumeric + `-`).",
            "example": "guess-uuid"
          }
        },
        "required": [
          "guess",
          "clientActionId"
        ]
      },
      "WordleFeedbackDto": {
        "type": "object",
        "properties": {
          "letter": {
            "type": "string",
            "example": "c"
          },
          "status": {
            "type": "string",
            "enum": [
              "correct",
              "present",
              "absent"
            ]
          }
        },
        "required": [
          "letter",
          "status"
        ]
      },
      "WordleGuessResponseDto": {
        "type": "object",
        "properties": {
          "guessCount": {
            "type": "number",
            "example": 1
          },
          "attemptsLeft": {
            "type": "number",
            "example": 5
          },
          "feedback": {
            "description": "Per-letter feedback.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WordleFeedbackDto"
            }
          },
          "wordleStatus": {
            "type": "string",
            "enum": [
              "IN_PROGRESS",
              "WON",
              "LOST",
              "TIMEOUT"
            ]
          }
        },
        "required": [
          "guessCount",
          "attemptsLeft",
          "feedback",
          "wordleStatus"
        ]
      },
      "FinishBodyDto": {
        "type": "object",
        "properties": {
          "clientActionId": {
            "type": "string",
            "description": "Idempotency key (8-64 chars, alphanumeric + `-`).",
            "example": "finish-uuid"
          }
        },
        "required": [
          "clientActionId"
        ]
      },
      "GameFinishResponseDto": {
        "type": "object",
        "properties": {
          "score": {
            "type": "number",
            "description": "Final score.",
            "example": 800
          },
          "timeMs": {
            "type": "number",
            "description": "Milliseconds from session start to finalization.",
            "example": 93211
          }
        },
        "required": [
          "score",
          "timeMs"
        ]
      },
      "ShadowQuestionPublicDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Question id."
          },
          "assetUrl": {
            "type": "string",
            "description": "CDN URL of the character shadow image."
          },
          "options": {
            "description": "Answer options — includes the correct one (not identified).",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "assetUrl",
          "options"
        ]
      },
      "ShadowStartResponseDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "Game session id."
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "Server-side expiry — the countdown is visual only."
          },
          "questions": {
            "description": "Round questions in order; correct answers are hidden.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShadowQuestionPublicDto"
            }
          },
          "maxAttemptsPerQuestion": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "sessionId",
          "expiresAt",
          "questions",
          "maxAttemptsPerQuestion"
        ]
      },
      "ShadowAnswerBodyDto": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "string",
            "description": "Question id from start.",
            "example": "cmspxk2y30006z963kwd5pf5a"
          },
          "answer": {
            "type": "string",
            "description": "Chosen option; matched case-insensitively.",
            "example": "Black Cat"
          },
          "clientActionId": {
            "type": "string",
            "description": "Idempotency key (8-64 chars, alphanumeric + `-`).",
            "example": "answer-uuid"
          }
        },
        "required": [
          "questionId",
          "answer",
          "clientActionId"
        ]
      },
      "ShadowAnswerResponseDto": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "string"
          },
          "questionIndex": {
            "type": "number",
            "example": 0
          },
          "correct": {
            "type": "boolean"
          },
          "correctAnswer": {
            "type": "string",
            "description": "Revealed only after the question resolves."
          },
          "attemptsUsed": {
            "type": "number",
            "example": 1
          },
          "attemptsLeft": {
            "type": "number",
            "example": 2
          },
          "questionScore": {
            "type": "number",
            "example": 100
          },
          "totalCorrect": {
            "type": "number",
            "example": 1
          },
          "questionCount": {
            "type": "number",
            "example": 6
          },
          "status": {
            "type": "string",
            "enum": [
              "IN_PROGRESS",
              "COMPLETED",
              "TIMEOUT"
            ]
          }
        },
        "required": [
          "questionId",
          "questionIndex",
          "correct",
          "correctAnswer",
          "attemptsUsed",
          "attemptsLeft",
          "questionScore",
          "totalCorrect",
          "questionCount",
          "status"
        ]
      },
      "CardPublicDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Card id (uuid).",
            "example": "a00c6eca-b097-430c-b799-b1fdcf6a22e1"
          },
          "index": {
            "type": "number",
            "description": "Position index on the 100-card board.",
            "example": 0
          }
        },
        "required": [
          "id",
          "index"
        ]
      },
      "CardsStartResponseDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "Game session id."
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "Server-side expiry — the countdown is visual only."
          },
          "cards": {
            "description": "Board positions (ids only — fronts are hidden).",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CardPublicDto"
            }
          },
          "backAssetUrl": {
            "type": "string",
            "description": "Shared card back asset URL."
          }
        },
        "required": [
          "sessionId",
          "expiresAt",
          "cards",
          "backAssetUrl"
        ]
      },
      "CardsMoveBodyDto": {
        "type": "object",
        "properties": {
          "cardId": {
            "type": "string",
            "description": "Card id (uuid) from start.",
            "example": "a00c6eca-b097-430c-b799-b1fdcf6a22e1"
          },
          "clientActionId": {
            "type": "string",
            "description": "Idempotency key (8-64 chars, alphanumeric + `-`).",
            "example": "move-uuid"
          }
        },
        "required": [
          "cardId",
          "clientActionId"
        ]
      },
...
      "CardsMoveResponseDto": {
        "type": "object",
        "properties": {
          "moveId": {
            "type": "string",
            "description": "The idempotency key echoed back."
          },
          "cardId": {
            "type": "string"
          },
          "frontAssetUrl": {
            "type": "string",
            "description": "Front asset — revealed only on this flip."
          },
          "revealed": {
            "type": "boolean",
            "description": "First card of an attempt is now face-up."
          },
          "matched": {
            "type": "boolean"
          },
          "matchCompleted": {
            "type": "boolean",
            "description": "This flip resolved a pair; both cards stay face-up."
          },
          "unmatchedFlipBack": {
            "type": "boolean",
            "description": "Attempt mismatched; both cards flip back."
          },
          "state": {
            "$ref": "#/components/schemas/CardsStatePublicDto"
          }
        },
        "required": [
          "moveId",
          "cardId",
          "frontAssetUrl",
          "revealed",
          "matched",
          "matchCompleted",
          "unmatchedFlipBack",
          "state"
        ]
      },
...
    }
  }
}